+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.90
 
 - changed to close xray only once per page load.
+- changed not skipping intros or recaps on episode 1, since they are often recaps of new seasons.
 
 ## 1.1.89